chore(app): 清理 Rust Clippy 告警并收紧 CI 门禁

+++ src/find/walker/dir_windows/nt_scan.rs
@@ -3,6 +3,7 @@ use super::eval::evaluate_entry_fast;
 use super::time::{filetime_ticks_to_secs, to_wide_null};
 use super::*;
 
+#[allow(clippy::too_many_arguments)]
 pub(super) fn scan_dir_nt(
     dir: &Path,
     base_root: &Path,
@@ -66,7 +67,7 @@ pub(super) fn scan_dir_nt(
         }
         restart = false;
 
-        let bytes = io_status.Information.min(buffer.len()) as usize;
+        let bytes = io_status.Information.min(buffer.len());
         let buffer_start = buffer.as_ptr() as usize;
         let buffer_end = buffer_start + bytes;
         let mut offset = 0usize;
